+++ libs/core/src/lib/dynamic-side-content/dynamic-side-content-side.component.spec.ts
@@ -1,0 +1,31 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+
+import { DynamicSideContentSideComponent } from './dynamic-side-content-side.component';
+import { CLASS_NAME } from './constants';
+
+describe('DynamicSideContentSideComponent', () => {
+    let fixture: ComponentFixture<DynamicSideContentSideComponent>;
+    let component: DynamicSideContentSideComponent;
+
+    beforeEach(async(() => {
+        TestBed.configureTestingModule({
+            imports: [CommonModule],
+            declarations: [DynamicSideContentSideComponent]
+        }).compileComponents();
+    }));
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(DynamicSideContentSideComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(fixture).toBeTruthy();
+    });
+
+    it('should add className to host', () => {
+        expect(fixture.debugElement.classes[CLASS_NAME.side]).toBeTrue();
+    });
+});