+++ libs/core/src/lib/layout-grid/directives/layout-grid-row.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    selector: '[fd-layout-grid-row], [fdLayoutGridRow]',
+    host: {
+        class: 'fd-row'
+    }
+})
+export class LayoutGridRowDirective {}