+++ libs/core/src/lib/product-switch/product-switch/product-switch.component.ts
@@ -1,6 +1,6 @@
-import { ChangeDetectionStrategy, Component, ViewEncapsulation, Input } from '@angular/core';
-import { PopoverComponent } from '../../popover/popover.component';
+import { ChangeDetectionStrategy, Component, ViewEncapsulation, Input, HostBinding } from '@angular/core';
 import {Placement} from 'popper.js';
+import { BasePopoverClass } from '../../popover/base/base-popover.class';
 
 @Component({
     selector: 'fd-product-switch',
@@ -9,12 +9,15 @@ import {Placement} from 'popper.js';
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProductSwitchComponent extends PopoverComponent {
-    @Input () 
+export class ProductSwitchComponent extends BasePopoverClass {
+    @Input ()
     placement: Placement = 'bottom-end'
-
 
     /**Input to set the aria label */
     @Input ()
     ariaLabel = 'popover-button';
+
+    /** Whether the product switch is disabled. */
+    @Input()
+    disabled = false;
 }