+++ libs/core/src/lib/utils/directives/carousel/carousel.module.ts
@@ -2,10 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CarouselItemDirective } from './carousel-item.directive';
 import { CarouselDirective } from './carousel.directive';
+import { CarouselComponent } from '../../../carousel/carousel.component';
+import { CarouselItemComponent } from '../../../carousel/carousel-item/carousel-item.component';
 
 @NgModule({
     imports: [CommonModule],
-    exports: [CarouselItemDirective, CarouselDirective],
-    declarations: [CarouselItemDirective, CarouselDirective]
+    exports: [CarouselItemDirective, CarouselDirective, CarouselComponent, CarouselItemComponent],
+    declarations: [CarouselItemDirective, CarouselDirective, CarouselComponent, CarouselItemComponent]
 })
 export class CarouselModule {}